requirements.txt
accelerate==0.30.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
airllm==2.6
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
async-timeout==5.0.1
asyncio==4.0.0
asyncpg==0.31.0
attrs==25.4.0
backcall==0.2.0
beautifulsoup4==4.14.3
bitsandbytes==0.43.1
blis==1.3.3
bs4==0.0.2
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
cryptography==46.0.5
cymem==2.0.13
datasets==2.19.1
decorator==5.2.1
dill==0.3.8
distro==1.9.0
docstring_parser==0.17.0
dotenv==0.9.9
einops==0.8.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.124.2
fastapi-shell==0.1.3
filelock==3.20.0
frozenlist==1.8.0
fsspec==2024.3.1
greenlet==3.3.0
groq==1.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
ipython==8.6.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
langdetect==1.0.9
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
murmurhash==1.0.15
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.1.105
ollama==0.6.1
optimum==1.13.2
packaging==25.0
pandas==2.3.3
parso==0.8.5
pdfminer.six==20251230
pdfplumber==0.11.9
peft==0.11.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==12.0.0
preshed==3.0.12
prompt-toolkit==3.0.30
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pyarrow-hotfix==0.7
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pypdfium2==5.6.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
redis==7.1.0
regex==2025.11.3
requests==2.32.5
rich==14.3.2
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
sniffio==1.3.1
soupsieve==2.8.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.45
srsly==2.5.2
stack-data==0.6.3
starlette==0.50.0
sympy==1.13.1
thinc==8.3.10
threadpoolctl==3.6.0
tokenizers==0.19.1
tomli==2.3.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
traitlets==5.14.3
transformers==4.43.3
triton==3.1.0
trl==0.9.4
typeguard==4.4.4
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tyro==1.0.6
tzdata==2025.2
urllib3==2.6.1
uvicorn==0.38.0
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.3
wrapt==2.1.2
xxhash==3.6.0
yarl==1.22.0
